v2v: Reduce amount of memory back down to around 2.5G
The amount of memory requested depends on libguestfs.  In this
libguestfs commit:
https://github.com/libguestfs/libguestfs/commit/224f373043302845122bf701ffc6e3416e0168fa
the size of the appliance was increased to 1280M.  This causes
virt-v2v to allocate 3.5G, which is much more than we need.

Previous virt-v2v commit fa8d5419c830eb942679d9e341b2b67f58324a31
adjusted the calculation to bring the total back down to 2G.  This
commit adjusts the calculation again, bringing it back down to 2.5G.