output: -o rhv: Change UID when creating the output disk
Virt-v2v 1.45 had this code in v2v/output_rhv.ml:

  method disk_create ?backingfile ?backingformat ?preallocation ?compat
    ?clustersize path format size =
    Changeuid.func changeuid_t (
      fun () ->
      ... code here for creating the disk ...

This was not replicated in virt-v2v 2.0 meaning that disks were not
created using the vdsm (36:36) user.  In some cases this would cause
disk creation to fail with an obscure error.  (For some reason
qemu-img does not print an error message in this case - the only way
to diagnose it is to strace the program).

The fix is to use the Changeuid wrapper when creating the disk.