convert: Fix --root ask so "exit" causes virt-v2v to exit
When the convert helper was split from modular virt-v2v I left this
case so that it called "exit 0".  However now that causes the parent
process virt-v2v to think that conversion was successful so it
proceeds with the copy.

Instead make the convert helper fail with an error.

I checked other places were helpers exit with code 0 and those are all
fine.