convert: Turn helper into an OCaml module
Splitting virt-v2v into separate helper programs was an interesting
exercise, but the helpers themselves were not independently useful
(eg. you could not run them on their own).  On reflection the value of
the exercise was really twofold:

 - providing the v2vdir and input and output pipelines based on NBD

 - forcing stages to be independent of each other (because they ran as
   separate processes), thus enforcing the split

The drawback of separate processes is:

 - need to serialise and then reparse all options as command line
   parameters

 - various smaller problems with getting accurate timing in messages,
   handling errors well, etc.

This commit turns helper-v2v-convert back into an internal module.  It
still has the strong separation, but will now be used directly as an
OCaml module.