Remove requested_guestcaps / rcaps
This was part of the old in-place support.  When we add new in-place
support we'll do something else, but currently this is dead code so
remove it completely.

Note this removes the code for installing the virtio-scsi driver (only
ever using virtio-blk).  This was also dead code in the current
implementation of virt-v2v, but worth remembering in case we want to
resurrect this feature in a future version.

Acked-by: Laszlo Ersek <lersek@redhat.com>