v2v: Force format of input to be specified
qemu 6.1 unnecessarily insists on the backing format of files being
set.  Change the type of the input disk so the format is no longer an
option, but must be set by the input mode.

This change is only required on the 1.44 branch, since modular
virt-v2v uses the qemu-nbd -s (snapshot) option to do the equivalent
which seems to handle the backing format automatically.

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=2025769
Reported-by: Vera Wu