input: -i vmx: Remove support for openssh scp < 8.8
Older openssh scp (< 8.8) had a strange double-quoting syntax for
paths because it passed the literal parameter to the remote side where
it was interpreted again by the remote shell.  In openssh 8.8 scp was
changed to use the sftp protocol instead.

Since we now require openssh >= 8.8, the -T option will always be
present so we don't need to test for it.

There are a few reasons why this change does not attempt to support
the older scp:

 * Distros like to keep up with OpenSSH for security reasons.

 * The older double quoting format was crazy, and probably insecure.
   Replacing it with a sane alternative is something we should
   encourage.

 * Import from -i vmx is fairly niche, and there are workable
   alternatives (VDDK or HTTPS) in case anyone is affected by this.

 * In fact it _does_ support the older version, at least in the case
   where the filename doesn't contain spaces or quoted characters.

Note that we require sftp in nbdkit-ssh-plugin, so we don't need to
worry about the remote server not supporting it because conversion
would fail anyway.

Reported-by: Ming Xie
Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=2027673