input: Split up the very large input/input.ml file
This file was contenated from the various input modes when I
modularized virt-v2v.  It was done this way essentially for
convenience.  However it wasn't a great idea because:

 - Not very easy to see which file is responsible for each mode.
   ie. it's a lot more obvious that input_disk.ml might be responsible
   for -i disk.

 - It didn't sufficiently isolate each mode.  Each mode is logically
   separate and does not need to be in the same top level file or
   module.

 - This single file "open"'d just about every namespace which is a
   symptom of poor encapsulation.

In addition the various *_source and *_servers functions which were
called one after another are combined into a single function.  (Their
separation was an unusual side effect of earlier modularization, not a
necessary feature).

This is just code movement, there is no change to functionality.