input: -it vddk: Reduce cow-block-size to 4K
NB: Requires the following fix to nbdkit:
https://gitlab.com/nbdkit/nbdkit/-/commit/c6fe9cb5b4a45e026f6c45ff3f92a510731d87b9

In commit 17358abd ("inputs: vddk: Use cow-block-size=1M") we started
to use a larger cow filter block size than the default.  This is based
on the theory that a larger block size causes larger, more efficient
read and extents requests to VDDK.  Some data were shown in the commit
message to show this was more efficient at the time.

However further testing shows it's not clear cut.  The trimming step
("Mapping filesystem data") is considerably slower with a larger block
size.  This is because the cow filter does a read-modify-write cycle
on the unaligned head and tail of the request (most trims and small
and/or not aligned to large block sizes), and the read part of this
cycle goes through VDDK which is slow.  The overall results are
clearly now better with a small block size, as shown below, even
though the copy step is slower, possibly because less data has been
read through the cow filter and cached during conversion.

Note that nbdkit-cow-filter coalesces operations so the original
rationale from that commit seems dubious.

Before (note time spent in "Mapping" step):

[   2.8] Opening the source
[  12.6] Inspecting the source
[  18.7] Checking for sufficient free disk space in the guest
[  18.7] Converting Windows 10 Enterprise to run on KVM
virt-v2v: This guest has virtio drivers installed.
[  36.1] Mapping filesystem data to avoid copying unused and blank areas
[  77.0] Closing the overlay
[  77.9] Assigning disks to buses
[  77.9] Checking if the guest needs BIOS or UEFI to boot
[  77.9] Copying disk 1/1
█ 100% [****************************************]
[ 332.3] Creating output metadata
[ 332.3] Finishing off

After:

[   2.8] Opening the source
[   7.9] Inspecting the source
[  12.8] Checking for sufficient free disk space in the guest
[  12.8] Converting Windows 10 Enterprise to run on KVM
virt-v2v: This guest has virtio drivers installed.
[  26.6] Mapping filesystem data to avoid copying unused and blank areas
[  28.0] Closing the overlay
[  28.7] Assigning disks to buses
[  28.7] Checking if the guest needs BIOS or UEFI to boot
[  28.7] Copying disk 1/1
█ 100% [****************************************]
[ 297.4] Creating output metadata
[ 297.4] Finishing off

Fixes: commit 17358abdcd44e3ff2a10fb16b3b626f438f2fc7a