input: Drop all use of nbdkit-readahead-filter
Previously we used nbdkit-readahead-filter with input modes that used
curl, ssh and vddk nbdkit plugins.  We quickly dropped it for curl and
vddk because it caused poor performance and failures with VMware
servers (bugs https://bugzilla.redhat.com/show_bug.cgi?id=1848862
https://bugzilla.redhat.com/show_bug.cgi?id=1832805).

This drops it for ssh too (eg for vmx or xen-ssh input modes).

The readahead filter in nbdkit is naively written and implementing a
replacement has proven difficult.  In any case it is not needed.  When
doing the conversion, readahead is already implemented in the
libguestfs appliance kernel.  When doing the copy, nbdcopy is already
able to keep sufficient requests in flight.