lib: Make Nbdkit mini-library handle mutable
Previously the Nbdkit.cmd handle was immutable, which is nice from a
purity point of view but made for a lot of boilerplate.  If we make it
mutable we can make the code a little smaller and easier to read and
write.

There is no significant change here and this is mostly refactoring.
However I also cleaned up the way environment variables are done
(LANG=C was duplicated before), and also the handling of the verbose
flag.