input: Remove nbdkit file fadvise=sequential option in a couple more places
The result of commit ac59d3b231 ("output: Don't use nbdkit-file-plugin
cache=none when writing") and the later partial revert in
commit 5d5383b912 is that we deliberately dropped fadvise=sequential
when using nbdkit-file-plugin, while keeping cache=none.

This flag was giving an incorrect hint to the kernel.  It was still
being used in a couple of places in virt-v2v, this commit drops those
too for the same reason.