convert_linux: translate the first CD-ROM's references in boot conf files
If the only CD-ROM in "s_removables" is on an IDE controller, and the
guest kernel represents it with a /dev/hdX device node, then convert
references to this device node, in the boot config files, to /dev/cdrom.
On the destination (after conversion), /dev/cdrom will point to whataver
node we converted the CD-ROM to, masking a potential i440fx -> q35 (IDE ->
SATA) board change.

If the only CD-ROM is not on an IDE controller, or the guest is modern
enough to represent the IDE CD-ROM as /dev/sr0, then perform no
translation. Namely, /dev/sr0 survives a potential i440fx -> q35 (IDE ->
SATA) board change intact.

When multiple CD-ROMs exist, emit a warning, and attempt the conversion on
the first CD-ROM, as a guess. This may be inexact, but we can't do better,
because:

- SATA, SCSI, and (on modern guests) IDE CD-ROMs are lumped together in
  the /dev/sr* namespace, on the source side, and "s_removable_slot" is
  useless for telling them apart, as we don't know the exact controller
  topology (and OS enumeration order);

- after conversion: some OSes don't create /dev/cdrom* symlinks to all
  CD-ROMs, and even if multiple such symlinks are created, their order is
  basically random.

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1637857
Signed-off-by: Laszlo Ersek <lersek@redhat.com>
Message-Id: <20211215153153.13468-1-lersek@redhat.com>
Acked-by: Richard W.M. Jones <rjones@redhat.com>