v2v: Swap over the output and conversion stages
In old virt-v2v, we did (approximately, since the steps were not as
clear):

  Input -> Convert -> Output -> Copy -> Finalize

After modularizing virt-v2v we changed this to:

  Input -> Output -> Convert -> Copy -> Finalize

However this has a (sort of) problem.  For -o rhv-upload when we start
the nbdkit rhv-upload-plugin machinery, it obtains a time-limited
ticket from imageio.  This ticket could expire if the conversion step
takes longer than a certain time (60 seconds by default, may be
increased in a future version of oVirt).

I believe this is really a problem in imageio or that the
rhv-upload-plugin should really renew this ticket automatically, but
it does not, instead failing.  (The ticket *is* renewed automatically
whenever a request is sent to imageio, but in this case no requests
are being sent).

Anyway the easiest thing is to switch the ordering back to how it was
in old virt-v2v (at top).  It doesn't make a difference for any other
output modes.

Reported-by: Nir Soffer