convert/libosinfo-c.c: turn caml_copy_*() return blocks into root values
caml_copy_string() and caml_copy_int64() create OCaml blocks
<https://ocaml.org/manual/intfc.html#sss:c-simple-allocation> that as such
are subject to movement or release by the garbage collector
<https://ocaml.org/manual/intfc.html#ss:c-blocks>.

In order for the GC to see that a block is in use (and should not be
collected, plus its inward references should be updated when it is moved),
the block needs to be reachable (directly or indirectly) from a "root"
value.

In each of the following patterns:

  Store_field (parent_block_v, field_n, caml_copy_string (str))
  Store_field (parent_block_v, field_n, caml_copy_int64 (i64))

the block produced by caml_copy_*() hangs in the air temporarily, before a
reference to it is added, originating from the parent block. If
Store_field() triggered a garbage collection just before creating the
reference, the retval block of caml_copy_*() could be moved (or even
released), and then the new reference inside the parent block would point
to garbage.

Store_field() used to be documented [*] to never invalidate its last
argument by calling the garbage collector, so the code we currently have
in "libosinfo-c.c" should be safe.

[*] The OCaml code comment

    [caml_modify] never calls the GC.

    disappeared in commits 5da7dd6d54d6 and 09c794232adb, on one of the
    branches between fork point commit bd7fa181cb64 and merge commit
    c8f58f406d57, and then the comment removal prevailed in the merge.
    OCaml has not had the comment reinstated since, with "trunk" currently
    at 284834d31767.

Either way, the following warning
<https://ocaml.org/manual/intfc.html#ss:c-simple-gc-harmony> does apply to
Store_field():

> The first argument of Store_field and Store_double_field must be a
> variable declared by CAMLparam* or a parameter declared by CAMLlocal* to
> ensure that a garbage collection triggered by the evaluation of the
> other arguments will not invalidate the first argument after it is
> computed.

meaning that the *first* argument of Store_field() *could* be released or
moved by a garbage collection that is triggered by the evaluation of the
last argument. And, in order to protect the first arg from such a release
(or, to be notified if the first arg is moved), the first arg must be
(temporarily) registered as a "root" value (with the "rootness"
substituting for other inward edge(s) from other blocks).

Keeping in mind the first and last (block) args' *different* exposures to
garbage collection is difficult. Therefore, for every Store_field() call
where the last arg is not an unboxed integer (in practice: Val_bool()),
but a block created with caml_copy_*(), track that block (at least
temporarily -- "long enough") as a root value, via a new local CAML value
called "copyv".

For consistency, update the one

  CAMLreturn (caml_copy_string (id))

call we have as well.

Suggested-by: Richard W.M. Jones <rjones@redhat.com>
Ref: https://listman.redhat.com/archives/libguestfs/2022-January/msg00037.html
Signed-off-by: Laszlo Ersek <lersek@redhat.com>
Message-Id: <20220110132623.9113-1-lersek@redhat.com>
Acked-by: Richard W.M. Jones <rjones@redhat.com>
[lersek@redhat.com: extend the commit message with the (rough) history of
 the caml_modify comment]