lib/utils.ml: Make with_nbd_connect_unix meta_contexts erasable
Make the meta_contexts parameter of Utils.with_nbd_connect_unix
optional and erasable.  This involves three connected changes, we
first make it optional (ie. '?meta_contexts'), providing the obvious
default of empty list.  We then need to move it to the first parameter
so OCaml can erase it even if we partially apply this function.
Finally remove the label from the function parameter 'f'.

Explanation for why the final change was necessary here:
https://listman.redhat.com/archives/libguestfs/2022-January/msg00082.html

Reviewed-by: Laszlo Ersek <lersek@redhat.com>