input: xen: Fix assertion error when importing from remote block device
We never supported this because OpenSSH sftp server does not know how
to open and get the size of a block device.  Before modular virt-v2v,
we would detect this situation and print an error:

  virt-v2v: error: guest disk sda appears to be zero bytes in size.

  There could be several reasons for this:

  Check that the guest doesn't really have a zero-sized disk.  virt-v2v
  cannot convert such a guest.

  If you are converting a guest from an ssh source and the guest has a disk
  on a block device (eg. on a host partition or host LVM LV), then
  conversions of this type are not supported.  See the virt-v2v-input-xen(1)
  manual for a workaround.

This error was lost in the conversion to modularity, but in any case
the proper way to detect this is in the input_xen driver itself.

In addition to this, when we removed the old virt-v2v-copy-to-local
tool, the text referring to the workaround in the manual became
meaningless, so I added the (ugly, manual) workaround that you have to
do back into the manual, minus the bit about the tool that we removed.

In future we could consider other ways to convert a remote block
device over SSH transparently.  (Sending a "dd" command, maybe?)

I also checked what happens when we try to convert an empty disk, and
it fails during inspection with a reasonable error message, so we
don't need to do anything to restore the above error:

  $ virt-v2v -i disk /var/tmp/empty -o null
  [   1.1] Opening the source
  [   7.0] Inspecting the source
  virt-v2v: error: inspection could not detect the source guest (or physical
  machine).

  Assuming that you are running virt-v2v/virt-p2v on a source which is
  supported (and not, for example, a blank disk), then this should not
  happen.

  No root device found in this operating system image.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=2041852
Reported-by: Xiaodai Wang