Restore message about setting up the input and output
Old virt-v2v would print a summary of the input and output options
before connecting to the input/output, looking something like this:

  [   0.2] Opening the source -i libvirt -ic [etc]

This gave reassurance that virt-v2v was doing something in the case
where the source was slow or unreachable.  In particular if you use
-i libvirt with a vCenter URL, and the URL is wrong, libvirt hangs for
a few minutes without printing anything.

Modular virt-v2v rearranged things so the connecting phase was silent,
which meant that in the case above virt-v2v appeared to hang for a few
minutes printing nothing at all.

This change adds to_string functions to all the input and output
methods and uses them to print a message like:

  [   0.0] Setting up the source: -i libvirt -ic [etc]

The hang still happens, but at least it's now clear where it's hanging.

Note the old "Opening the source" message now refers to libguestfs
connecting to the NBD source disk pipeline.

Typical full output looks like this:

  $ virt-v2v -i disk /var/tmp/fedora-35.img -o disk -os /var/tmp/out
  [   0.0] Setting up the source: -i disk /var/tmp/fedora-35.img
  [   1.1] Opening the source
  [   5.9] Inspecting the source
  [  11.5] Checking for sufficient free disk space in the guest
  [  11.5] Converting Fedora Linux 35 (Thirty Five) to run on KVM
  virt-v2v: warning: /files/boot/grub2/device.map/hd0 references unknown
  device "vda".  You may have to fix this entry manually after conversion.
  virt-v2v: This guest has virtio drivers installed.
  [  57.4] Mapping filesystem data to avoid copying unused and blank areas
  [  61.0] Closing the overlay
  [  61.7] Assigning disks to buses
  [  61.7] Checking if the guest needs BIOS or UEFI to boot
  [  61.7] Setting up the destination: -o disk -os /var/tmp/out
  [  62.8] Copying disk 1/1
  █ 100% [****************************************]
  [  81.7] Creating output metadata
  [  81.7] Finishing off

Reported-by: Xiaodai Wang
Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=2041886
Fixes: commit 255722cbf39afc0b012e2ac00d16fa6ba2f8c21f
Reviewed-by: Laszlo Ersek <lersek@redhat.com>