tests: Don't use virt-builder --update when testing Windows conversions
The error was:

virt-builder: error: cannot use ‘--update’ because no package manager
has been detected for this guest OS.

If this guest OS is a common one with ordinary package management then this
may have been caused by a failure of libguestfs inspection.

For OSes such as Windows that lack package management, this is not
possible.  Try using one of the ‘--firstboot*’ flags instead (described
in the manual).

Reviewed-by: Laszlo Ersek <lersek@redhat.com>