lib/qemuNBD.ml: Use qemu-nbd --shared=0 flag to allow multiple connections
qemu-nbd --shared (-e) flag controls how many clients can connect
concurrently.  0 means unlimited.

We want to allow the sockets to be queried by other processes while
virt-v2v is running and it should be safe to do this.  The default
configuration of qemu-nbd doesn't allow this so add --shared=0.

Note this does not (in current qemu) enable multi-conn because we
aren't using the -r (read-only) flag.

Reviewed-by: Laszlo Ersek <lersek@redhat.com>