input: libvirt: Share a single connection to the source NBD server
When using virt-p2v from RHEL 7, it starts a very old qemu-nbd server
(probably 1.5.3) which required the --shared parameter to enable
sharing even in read-only mode.  Since it doesn't pass this parameter
only a single connection at a time is allowed, and further connections
will deadlock.  Note that later versions of qemu-nbd changed this so
that read-only connections permit sharing automatically.

In modular virt-v2v we now use nbdkit-nbd-plugin to proxy the
connection to virt-p2v / qemu-nbd.  When you connect to this multiple
times, as virt-v2v does, it will make multiple connections to the
backend qemu-nbd.  This will cause a deadlock.

We can use the nbdkit-nbd-plugin shared=true flag to enable the plugin
to share a single connection to the backend between multiple nbdkit
clients.

Reported-by: Tingting Zheng
Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=2044911
Reviewed-by: Laszlo Ersek <lersek@redhat.com>