output: -o rhv-upload: Kill nbdkit instances before finalizing
If nbdkit instance(s) are still running then they will hold open some
idle http connections to imageio.  In some versions of imageio,
starting finalization in this state causes a 60 second delay.

This patch kills the nbdkit instances between RHVUpload.setup and
RHVUpload.finalize.  It's not very nice to do this because it leaves
hanging outX sockets, but we can revert this patch in future when we
can assume new enough imageio is installed on the target.

For more details see this thread:
https://listman.redhat.com/archives/libguestfs/2022-February/thread.html#00109

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=2039255
Thanks: Nir Soffer, Laszlo Ersek
Reviewed-by: Laszlo Ersek <lersek@redhat.com>