input: Refactor code in the input modules
Cleans up things left over from modularising virt-v2v and aims to make
the code more straightforward to follow.

In particular (just because of the incremental steps that
modularization went through) the input modules all looked like:

  let foo_source dir options args =
    ...
  module Foo = struct
    let setup dir options args = foo_source dir options args
  end

(and the same for other methods).  One of the refactorings is to
simply place the code directly inside the module functions.

I also moved around the module methods to reflect more closely the
order in which the code is actually called by virt-v2v.

The change looks large, but that's because of whitespace changes
because of reindenting the code.  There are no code changes here and
the refactoring should be completely neutral.