Refactor calculation of output name
When we decide which output name to use, we start with the source name
(source.s_name), and override it with the output name specified on the
command line (-on option).

This calculation is done in quite a few places.

One place is in the conversion function.  This is completely
unnecessary and the commit removes it.

Also all the output functions do it, but they don't do it during other
option parsing code (ie. parse_options).  I have rationalized this
slightly by moving the calculation in all output modules to the
parse_options function and passing the computed value around in the
same structure as all the other parsed options.  (This change will
make more sense in the context of the following commit which further
rationalizes option parsing across all output modules.)

The change is quite large but is only refactoring and should make no
semantic change.