convert_windows: install "v2vnetcf.ps1" with prio 2500
The windows converter installs the firstboot scripts in this order:

  do_convert
    configure_firstboot
      configure_wait_pnp
        add_firstboot_script
      configure_vmdp
        add_firstboot_script
      configure_qemu_ga
        add_firstboot_script
      unconfigure_xenpv
        add_firstboot_script
      unconfigure_prltools
        add_firstboot_script
      unconfigure_vmwaretools
        add_firstboot_script
    configure_network_interfaces
      install_firstboot_powershell "v2vnetcf.ps1"
        add_firstboot_script
          -> the script that polls for "netkvm.sys" with Get-NetAdapter,
             then calls New-NetIPAddress

Because of this, if the guest is supposed to have static IP address(es),
they are set by the "v2vnetcf.ps1" firstboot script quite late during
first boot.

It's more comfortable for users if the network is usable from an earlier
point of the first boot, so register "v2vnetcf.ps1" in priority group 2500
(the default is 5000), moving the script to the front.

Suggested-by: Richard W.M. Jones <rjones@redhat.com>
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1788823
Signed-off-by: Laszlo Ersek <lersek@redhat.com>
Message-Id: <20220308143059.6384-13-lersek@redhat.com>
Acked-by: Richard W.M. Jones <rjones@redhat.com>