input: Add options.read_only flag
This flag can be used for certain inputs to control whether we place a
protective overlay over the source.  It is only supported for local
file type inputs (-i disk and -i libvirt).  For other input types it
cannot work:

 * -i ova: These cannot be updated in-place because the OVA embeds
   checksums.

 * Any input on VMware or Xen: There's no point doing an in-place
   update for a guest which his still sitting on a foreign hypervisor.

 * -i vmx: We cannot reliably update a disk in VMDK format using qemu
   tools.

This flag is always set to true, so there is no effect in the current
code.  It allows us to implement a second front end for in-place
support.