v2v: Re-enable v2v_unit_tests
This test was disabled in commit 255722cbf3 ("v2v: Modular virt-v2v").
This simply enables it again.  I don't think the test is very valuable
so maybe we should completely remove it in future, but since the code
was still there and enabling it was trouble-free that's what I did.

Fixes: commit 255722cbf39afc0b012e2ac00d16fa6ba2f8c21f