docs: Mention pnp_wait.exe and refine pvvxsvc.exe documentation
Virt-v2v can use pnp_wait.exe to wait for new Windows devices to
activate.  However this was not mentioned in the existing
documentation.

Also refine the documentation for pvvxsvc.exe (used only by SUSE
hosts) so that we don't mention the non-existent --firstboot* options.
This was incorrect documentation copied from virt-customize.  Just say
that it's an alternative to RHSrvAny.

Reported-by: Ming Xie
Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=2059287#c6
Fixes: commit ee78c06b7f ("customize: add support for pvvxsvc")
Reviewed-by: Laszlo Ersek <lersek@redhat.com>