convert/convert_linux.ml: Refactor source device name calculation
This neutral refactoring changes the existing source device name
calculation so that it can handle (in the next commit) NVMe device
names.

Existing source device names such as "hda" or "sda" are composed of a
prefix ("hd", "sd") + alphabetic device index ("a").  NVMe devices
have a namespace after the prefix and use a numeric index ("nvme0n1"
== "nvme" + "0" + "n1").  By refactoring the calculation into a
function we will be able to handle both.

Reviewed-by: Laszlo Ersek <lersek@redhat.com>