docs: Remove documentation for -o json in preparation for removal
I would like to remove the -o json option in virt-v2v 2.2.

It was added to support KubeVirt, but never actually used.  It is
better for specific targets to implement a proper output mode (as we
will eventually do for KubeVirt) rather than relying on a custom
templating language which knows nothing about the target.

This preparatory patch removes only the documentation, although it
leaves a note in the main manual page that the feature existed and is
deprecated.

Reviewed-by: Laszlo Ersek <lersek@redhat.com>
(cherry picked from commit 02e58d54fef14be72078559dae8c7c7a39f7c0f5)