run: Use GLIBC_TUNABLES instead of MALLOC_PERTURB_
glibc 2.34 removed MALLOC_CHECK_ and MALLOC_PERTURB_ and replaced them
with GLIBC_TUNABLES with a completely different syntax and requiring a
special library (libc_malloc_debug.so.0) to be preloaded.

Note this works even if the preloaded library is not present since
ld.so ignores missing LD_PRELOAD.  So this shouldn't break anything
either for older glibc or if the special library is missing.

(cherry picked from commit 9f54a4063ddcda91a7bb965c343178da419a8726)