lib: Add a simple YAML generator
This is only enough to be able to generate Kubevirt machine
descriptions so it does not have to deal with all the general
complexity of YAML.  Only associative arrays, lists, simple types like
strings, and blocks of lines are supported.

Unfortunately even generating simple YAML safely is not easy.  There
are many quirks to the format and many traps, including the dependency
between lists and associative arrays, and the lack of type safety.